dvdplayer: add a clock sync stream type
This can be used to let input stream control the playback speed of dvdplayer
and for example keep multiple xbmc instances in synchronous playback.

Currently a psuevdo codec called BXA is used (based on the airtunes stream)
which passes a CurrentHostCounter() tick at which point the pts of the
packet is supposed to be heard/seen.